requirements.txt
flask
pandas
pillow
openpyxl
gunicorn
